reformat aws tests
